+++ src/core/abstract/SettingsHandler.ts
@@ -1,11 +1,29 @@
+import { Alert } from '@core/shared/Alert';
+import { Workspace } from '@core/shared/Workspace';
 import * as vscode from 'vscode';
-
 export abstract class SettingsHandler {
   constructor(
     private readonly settingsAliasKey: string | undefined = undefined,
   ) {}
 
+  protected static readonly settingsFileName = 'settings.json';
+
+  public resetToDefaults = async (): Promise<void> => {
+    throw new Error('Method is not implemented');
+  };
+
+  private isSettingsOpen = () => {
+    const { settingsFileName } = SettingsHandler;
+    return Workspace.getTextDocuments().find(
+      (doc) => doc.fileName.endsWith(settingsFileName) && doc.isDirty,
+    );
+  };
+
   protected updateConfig = async (sectionToUpdate: string, value: any) => {
+    if (this.isSettingsOpen()) {
+      Alert.warn(`Please save the ${SettingsHandler.settingsFileName} file.`);
+    }
+
     const config = vscode.workspace.getConfiguration(this.settingsAliasKey);
     const configTarget = vscode.ConfigurationTarget.Global;
     await config.update(sectionToUpdate, value, configTarget);
@@ -17,6 +35,11 @@ export abstract class SettingsHandler {
   ): T => {
     const config = vscode.workspace.getConfiguration(this.settingsAliasKey);
     const inspected = config.inspect<T>(sectionToInspect);
-    return inspected?.globalValue || defaultValue;
+
+    if (inspected?.globalValue === undefined) {
+      return defaultValue;
+    }
+
+    return inspected?.globalValue;
   };
 }