+++ src/core/settings/index.ts
@@ -1,11 +1,14 @@
-import { CSettings } from './c';
-import { CppSettings } from './cpp';
-import { ExtensionSettings } from './extension';
+import { CSettings } from './CSettings';
+import { EditorSettings } from './EditorSettings';
+import { ExtensionSettings } from './ExtensionSettings';
+import { MsCppSettings } from './MsCppSettings';
 
 export class Settings {
-  static cpp = new CppSettings();
+  static msCpp = new MsCppSettings();
 
   static ext = new ExtensionSettings();
 
   static c = new CSettings();
+
+  static editor = new EditorSettings();
 }