+++ src/core/shared/Alert.ts
@@ -5,6 +5,10 @@ export class Alert {
     return vscode.window.showInformationMessage(msg);
   };
 
+  static warn = (msg: string) => {
+    return vscode.window.showWarningMessage(msg);
+  };
+
   static error = (msg: string) => {
     return vscode.window.showErrorMessage(msg);
   };