+++ src/core/shared/Setup.ts
@@ -1,67 +1,83 @@
+import { changeCompilerProfile } from '@commands/changeCompilerProfile';
 import { compile8051 } from '@commands/compile8051';
 import { compilePc } from '@commands/compilePc';
+import { compilePic } from '@commands/compilePic';
 import { execPc } from '@commands/execPc';
 import { openKeybindings } from '@commands/openKeybindings';
 import { openLink } from '@commands/openLink';
 import { openSettings } from '@commands/openSettings';
 import { Constants } from '@core/Constants';
 import { GlobalWebViewProvider } from '@core/GlobalWebViewProvider';
+import { Settings } from '@core/settings';
 import { SyncSettings } from '@core/SyncSettings';
 import * as vscode from 'vscode';
 
+import { Context } from './Context';
 import { Keybindings } from './Keybindings';
 import { Logger } from './Logger';
 
 export class Setup {
-  static onActivate = (context: vscode.ExtensionContext) => {
+  private static compilerView: GlobalWebViewProvider;
+  private static generalView: GlobalWebViewProvider;
+
+  static onActivate = async (context: vscode.ExtensionContext) => {
     Logger.log(`Starting ${Constants.EXT_NAME}...`);
 
-    this.createViews(context);
-    this.registerCommands(context);
-    SyncSettings.startSync();
+    SyncSettings.addSyncListenner(this.syncSettingsWithWebview);
+    Context.setContext(context);
+    this.registerCommands();
+    this.createViews();
+    await SyncSettings.startSync();
+    await Context.afterActivation();
 
     Logger.log(`Started ${Constants.EXT_NAME}!`);
   };
 
-  private static registerCommands = (context: vscode.ExtensionContext) => {
-    const compile8051Command = vscode.commands.registerCommand(
-      'cimol-micro-intellisense.compiler.8051',
-      compile8051.exec,
-    );
-
-    const compilePcCommand = vscode.commands.registerCommand(
-      'cimol-micro-intellisense.compiler.pc',
-      compilePc.exec,
-    );
-
-    const execPcCommand = vscode.commands.registerCommand(
-      'cimol-micro-intellisense.compiler.runPc',
-      execPc.exec,
-    );
-
-    context.subscriptions.push(
-      compile8051Command,
-      compilePcCommand,
-      execPcCommand,
-    );
+  private static registerCommands = () => {
+    compile8051.registerInVsCode();
+    compilePc.registerInVsCode();
+    execPc.registerInVsCode();
+    compilePic.registerInVsCode();
+    changeCompilerProfile.registerInVsCode();
   };
 
-  private static createViews = (context: vscode.ExtensionContext) => {
-    const generalView = new GlobalWebViewProvider(context, 'generalView');
-    const compilerView = new GlobalWebViewProvider(context, 'compilerView');
+  private static syncSettingsWithWebview = async () => {
+    const currentProfile = Settings.ext.getProfile();
 
-    generalView.setWebViewListenners(openLink, openSettings, openKeybindings);
-    compilerView.setWebViewListenners(compile8051, compilePc, execPc);
-
-    context.subscriptions.push(generalView.register());
-    context.subscriptions.push(compilerView.register());
-
-    compilerView.onRender(() => {
-      compilerView.postMessage({
-        title: 'Supply commands keybindings',
-        command: 'supplyKeybindings',
-        payload: Keybindings.list(),
-      });
+    this.compilerView.postMessage({
+      title: 'Supply current compiler profile',
+      command: 'supplyCurrentProfile',
+      payload: currentProfile,
     });
+
+    this.compilerView.postMessage({
+      title: 'Supply commands keybindings',
+      command: 'supplyKeybindings',
+      payload: Keybindings.list(),
+    });
+  };
+
+  private static createViews = () => {
+    this.generalView = new GlobalWebViewProvider('generalView');
+    this.compilerView = new GlobalWebViewProvider('compilerView');
+
+    this.generalView.setWebViewListenners(
+      openLink,
+      openSettings,
+      openKeybindings,
+    );
+
+    this.compilerView.setWebViewListenners(
+      compile8051,
+      compilePc,
+      compilePic,
+      execPc,
+      changeCompilerProfile,
+    );
+
+    Context.registerSubscription(this.generalView.register());
+    Context.registerSubscription(this.compilerView.register());
+
+    this.compilerView.onRender(this.syncSettingsWithWebview);
   };
 }