+++ src/webviews/CompilerSidebar/ActionButton/index.tsx
@@ -1,27 +1,30 @@
 import styles from './styles.module.css';
 
-interface SidebarActionButtonProps {
-  actionIcon: React.ReactNode;
-  infoIcon: React.ReactNode;
-  label: string;
+interface ActionButtonProps {
+  icon: React.ReactNode;
   tip?: string;
-  onActionClick: () => void;
+  tipPos?: 'top' | 'bottom';
+  isDisabled?: boolean;
+  onClick: () => void;
 }
 
-export const ActionButton = (props: SidebarActionButtonProps) => {
-  const { actionIcon, infoIcon, onActionClick, label, tip = '' } = props;
+export const ActionButton = (props: ActionButtonProps) => {
+  const { icon, onClick, tip = '', isDisabled = false, tipPos = 'top' } = props;
+
+  const tipClass = `${styles.actionButtonTip} ${
+    tipPos === 'top'
+      ? styles['actionButtonTip--top']
+      : styles['actionButtonTip--bottom']
+  }`;
+
+  const buttonClass = `${styles.actionButton} ${
+    isDisabled ? styles['actionButton--disabled'] : {}
+  }`;
 
   return (
-    <li className={styles.compilerSidebarActionButton}>
-      <div className={styles.description}>
-        <div className={styles.infoIconContainer}>{infoIcon}</div>
-
-        <span className={styles.descriptionText}>{label}</span>
-      </div>
-      <button className={styles.button} onClick={onActionClick}>
-        {tip && <span className={styles.buttonTip}>{tip}</span>}
-        {actionIcon}
-      </button>
-    </li>
+    <button disabled={isDisabled} className={buttonClass} onClick={onClick}>
+      {tip && <span className={tipClass}>{tip}</span>}
+      {icon}
+    </button>
   );
 };